feat: start building mobile search
